Host icon on GitHub
I don't know why I didn't do this in the first place. Imgur doesn't lose images often, but GitHub is a less volatile host. 

This update is functionally the same as the initial release. 

Note: The icon will not work until this is merged from dev to main, as the URL is referencing the file that will be in Main upon merge.